// apps/chataigne/ui/src/lib/systems/state_machine/generated/ANodeOutputPreviewSampleDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ContextKeyDto } from "./ContextKeyDto";
import type { OutputPreviewStatusDto } from "./OutputPreviewStatusDto";
import type { RuntimeValueDto } from "./RuntimeValueDto";

export type ANodeOutputPreviewSampleDto = { formula_id: string, processor_id: string | null, context_key: ContextKeyDto | null, node_id: string, exec_node_id: string, output_socket_id: string, value_type: string, value: RuntimeValueDto, logical_tick: bigint, status: OutputPreviewStatusDto, };